// src/api/models/document-resource.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DocumentStatus } from '../models/document-status';
export interface DocumentResource {
  created_at: (string | null);
  description: (string | null);
  file: string;
  id: number;
  name: string;
  status: DocumentStatus;
  updated_at: (string | null);
}